GitAPIManager/requirements.txt
django >=1.9, <2
pygithub >=1.29
